fixed bugs refactored code added more cases for testing

+++ UIPerformance/Models/ResultModel.cs
@@ -4,6 +4,6 @@ namespace UIPerformance.Models
     {
         public ViewType Type { get; set; }
         public int ElapsedTime { get; set; }
-        public long ElapsedMemory { get; set; }
+        public decimal ElapsedMemory { get; set; }
     }
 }
